[feat] Implement pool sync methods of blocker
Among three major roles of blocker, this commit contains an
implementation of PullRequestPoolSync module.

It synchronizes all open pull requests for each IntegrationConfig and
store it to a cache.

Also, if a PR meets every conditions of mergeQuery, the PR is added to a
MergePool, which is a set of pR candidates to be merged.