[fix] Differentiate webhook sender and pull request author
In the case of pull-request edit, such as labelling and unlabelling,
webhook's sender is different from the pull request's author.

To authorize labelling/unlabelling event (#226), we should know the
exact sender (who actually labelled the pull request), not the author of
the pull request.

This commit differentiates webhook sedner and pull request author. In
addition, authors of issue comments are differentiated as well.