[fix] Invalidate pr pool cache if new commit is added (#230)
Before this commit, when a new commit is added to the pull request,
commit status of blocker to the pull request is not set until conditions
change, because the cache is treated 'not dirty'.

This commit fixes the issue, by setting the cache is dirty when the SHA
changes for the pull request.