[feat] Enable merge commit template (#275)
close #265

If commit template is specified in the IntegrationConfig's mergeConfig,
the template is compiled using blocker.PullRequest struct.
Before compiling it, commits of the pull request is parsed and set to
the structure's Commit field.
User can use the PR's info and commits for writing the message temaplte.

Also, fake git client's MergePullRequest method is implemented. As a
result, there are some modifications in the test codes.